Simplify accessing assembly versions
Use the same method everywhere an assembly version is required in order not to use`FileVersionInfo.GetVersionInfo` directly, which is not compatible with single file applications whose assembly.Location always return an empty string. (It was still used directly in the `AuthProcessor` class)

Also, use GetCustomAttribute<AssemblyFileVersionAttribute>()?.Version with a question mark before acessing the `.Version` property in order to fallback to version on disk rather than crashing with a null reference exception. (The ? was missing in both Microsoft.PowerPlatform.Dataverse.Client.ServiceClient.SdkVersionProperty and Microsoft.PowerPlatform.Dataverse.Client.Auth.TokenCache.FileBackedTokenCacheHints.FileBackedTokenCacheHints)