Actually parse the "Integrated Security" connection string part
Assign the result of parsing the "Integrated Security" setting to the `DataverseConnectionStringProcessor.UseCurrentUser` property. The "Integrated Security" part was parsed but the result was not used and the `UseCurrentUser` existed but was never assigned. I'm pretty sure that it was intended to be like this, especially with the comment above parsing of the IntegratedSecurity:
> // Check to see if use current user is configured.